Addition of Mood Based Spotify Playlist Finder
This project aims to find the most popular playlists out there from spotify using the SpotiPY library and Spotify Web API, based on the entered mood of the user.

The user is given a selected choice of mood to avoid inconsistency with the aim of this project.

After the user selects a mood, they can find suitable playlist link and the names of all the songs along with artist shown in a systematic manner